--- src/main/java/com/example/umc7th/article/dto/ArticleResponseDTO.java
@@ -1,22 +1,0 @@
-package com.example.umc7th.article.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ArticleResponseDTO {
-
-    private Long id;
-    private String title;
-    private String content;
-    private int likeNum;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-}
