--- src/main/java/com/example/umc7th/article/service/command/ArticleCommandService.java
@@ -1,8 +1,0 @@
-package com.example.umc7th.article.service.command;
-
-import com.example.umc7th.article.dto.ArticleRequestDTO;
-import com.example.umc7th.article.entity.Article;
-
-public interface ArticleCommandService {
-    Article createArticle(ArticleRequestDTO.CreateArticleDTO dto);
-}
