--- src/main/java/com/example/umc7th/article/service/command/ArticleCommandServiceImpl.java
@@ -1,24 +1,0 @@
-package com.example.umc7th.article.service.command;
-
-import com.example.umc7th.article.converter.ArticleConverter;
-import com.example.umc7th.article.dto.ArticleRequestDTO;
-import com.example.umc7th.article.entity.Article;
-import com.example.umc7th.article.repository.ArticleRepository;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-// 모든 메소드가 하나의 Transaction 단위로 동작, 단일 메소드에도 선언 가능
-@Transactional
-@RequiredArgsConstructor
-public class ArticleCommandServiceImpl implements ArticleCommandService{
-
-    private final ArticleRepository articleRepository;
-
-    @Override
-    public Article createArticle(ArticleRequestDTO.CreateArticleDTO dto) {
-        Article article = ArticleConverter.toEntity(dto);
-        return articleRepository.save(article);
-    }
-}
