+++ src/main/java/com/example/umc7th/domain/article/service/command/ArticleCommandService.java
@@ -1,0 +1,8 @@
+package com.example.umc7th.domain.article.service.command;
+
+import com.example.umc7th.domain.article.dto.ArticleRequestDTO;
+import com.example.umc7th.domain.article.entity.Article;
+
+public interface ArticleCommandService {
+    Article createArticle(ArticleRequestDTO.CreateArticleDTO dto);
+}
