+++ src/main/java/com/example/umc7th/domain/article/service/query/ArticleQueryService.java
@@ -1,0 +1,10 @@
+package com.example.umc7th.domain.article.service.query;
+
+import com.example.umc7th.domain.article.dto.ArticleResponseDTO;
+
+import java.util.List;
+
+public interface ArticleQueryService {
+    ArticleResponseDTO getArticle(Long id);
+    List<ArticleResponseDTO> getArticles();
+}
