+++ src/main/java/com/example/umc7th/domain/reply/dto/ReplyRequestDTO.java
@@ -1,0 +1,12 @@
+package com.example.umc7th.domain.reply.dto;
+
+import lombok.Getter;
+
+public class ReplyRequestDTO {
+
+    @Getter
+    public static class CreateReplyDTO {
+        private String content;
+    }
+
+}
