+++ src/main/java/com/example/umc7th/domain/reply/dto/ReplyResponseDTO.java
@@ -1,0 +1,20 @@
+package com.example.umc7th.domain.reply.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ReplyResponseDTO {
+
+    private Long id;
+    private String content;
+    private LocalDateTime createdAt;
+    private LocalDateTime updatedAt;
+}
