+++ src/main/java/com/example/umc7th/domain/reply/service/query/ReplyQueryService.java
@@ -1,0 +1,10 @@
+package com.example.umc7th.domain.reply.service.query;
+
+import com.example.umc7th.domain.reply.dto.ReplyResponseDTO;
+
+import java.util.List;
+
+public interface ReplyQueryService {
+    ReplyResponseDTO getReply(Long id);
+    List<ReplyResponseDTO> getReplies();
+}
