--- src/main/java/com/example/umc7th/reply/dto/ReplyResponseDTO.java
@@ -1,20 +1,0 @@
-package com.example.umc7th.reply.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ReplyResponseDTO {
-
-    private Long id;
-    private String content;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-}
