--- src/main/java/com/example/umc7th/reply/repository/ReplyRepository.java
@@ -1,7 +1,0 @@
-package com.example.umc7th.reply.repository;
-
-import com.example.umc7th.reply.entity.Reply;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReplyRepository extends JpaRepository<Reply, Long> {
-}
