--- src/main/java/com/example/umc7th/reply/service/command/ReplyCommandServiceImpl.java
@@ -1,30 +1,0 @@
-package com.example.umc7th.reply.service.command;
-
-import com.example.umc7th.article.entity.Article;
-import com.example.umc7th.article.repository.ArticleRepository;
-import com.example.umc7th.global.apiPayload.code.GeneralErrorCode;
-import com.example.umc7th.global.apiPayload.exception.GeneralException;
-import com.example.umc7th.reply.converter.ReplyConverter;
-import com.example.umc7th.reply.dto.ReplyRequestDTO;
-import com.example.umc7th.reply.entity.Reply;
-import com.example.umc7th.reply.repository.ReplyRepository;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@Transactional
-@RequiredArgsConstructor
-public class ReplyCommandServiceImpl implements ReplyCommandService {
-
-    private final ReplyRepository replyRepository;
-    private final ArticleRepository articleRepository;
-
-    @Override
-    public Reply createReply(ReplyRequestDTO.CreateReplyDTO dto, Long articleId) {
-        Article article = articleRepository.findById(articleId).orElseThrow(() -> new GeneralException(GeneralErrorCode.ARTICLE_NOT_FOUND_404));
-        // Convert는 Service 영역으로 통일
-        Reply reply = ReplyConverter.toEntity(dto, article);
-        return replyRepository.save(reply);
-    }
-}
